rearrange configuration
